Add ability for interceptors to handle exceptions in lacinia
This change enables the use of standard pedestal error handling for errors
raised in the interceptor stack or in lacinia itself. It does so by moving
response error generation to the very end of the (outbound) interceptor stack.
End users can then inject interceptors prior to it to handle or log exceptions
as desired.

As it stands, this is only applicable to the synchronous query executor. I
would also like to invoke the same error interceptors with exceptions
occurring on the asynchronous path, but am currently unable to figure out how
exceptions occurriing in channels should be handled in pedestal.

If we do figure that out, we can simplify some of the code here as there are
now two error paths (one for sync, one for async).